refactor: controller, domain, view로 패키징 분리
