+++ src/main/java/lottery/domain/GenerateCount.java
@@ -1,4 +1,4 @@
-package lottery;
+package lottery.domain;
 
 public class GenerateCount {
