+++ src/main/java/lottery/domain/Lottery.java
@@ -1,4 +1,4 @@
-package lottery;
+package lottery.domain;
 
 public class Lottery {
