+++ src/main/java/lottery/domain/LotteryNumber.java
@@ -1,4 +1,4 @@
-package lottery;
+package lottery.domain;
 
 import java.util.Objects;
