+++ src/test/java/lottery/domain/GenerateCountTest.java
@@ -1,5 +1,6 @@
-package lottery;
+package lottery.domain;
 
+import lottery.domain.GenerateCount;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
