+++ src/test/java/lottery/domain/LotteriesTest.java
@@ -1,4 +1,4 @@
-package lottery;
+package lottery.domain;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
